requirements.txt
dj-database-url==0.5.0
Django==4.0.5
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-shopping-cart==0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docker==5.0.3
docker-compose==1.29.2
gunicorn==20.1.0
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.9.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3.1
razorpay==1.3.0
whitenoise==6.1.0